[MOD - 6778] AVX512_FP16 distance functions (#477)
* change avx512 to avx512f to set the ground for avx512_fp16 (which is already used for avx512f for fp16)

* format

* more fixes

* add OPT_AVX512_FP16 macro if supported

IP & L2 : choose OPT_AVX512_FP16 if supported

add unit tests for fp16 with advanced flag (seperate test as the baseline calculation is different)
fp16 Benchmark:
all of these changes will be moved to a new PR
Duplicated bm class and made it templated only for fp16 so we can control the conversion function in SetUp
also added high dimension bm to fp16

* format

* install gcc12 ubuntu22

* set gcc12 as default

* merge with the bm spaces small refactor

* choose OPT_AVX512_FP16 depending on the dim
IP: 500
L2: 440

* Upgrade gcc in rocky 8 (12) + 9 (13)

* enable flow temp

* rocky8 and ubuntu22: install gcc 13

* fix temp flow

* jammy: revedrt to gcc12

* align gcov version with gcc version
temp flow runcodcov on focal (before updating gcov version)

* ubuntu20 align gcov version with gcc version

* fix cast to avoid warning

* add test for fp16 native type comparing to vecsim

allow 1% error in fp16 advanced tests

add explicit to float16 constructor to protected from implicit casts.

* fix comment
disable flow temp

* use explicit ctor for bfloat to avoid implicit casts

* Apply suggestions from code review


* use explicit ctor for bfloat to avoid implicit casts

* IP: mask mul instead of mask load as it is faster

---------

+++ src/VecSim/spaces/L2/L2_SSE3_BF16.h
@@ -57,7 +57,7 @@ float BF16_L2SqrSIMD32_SSE3(const void *pVect1v, const void *pVect2v, size_t dim
     // Handle residual % 4
     if constexpr (residual % 4) {
         __m128i v1, v2;
-        constexpr bfloat16 zero = 0;
+        constexpr bfloat16 zero = bfloat16(0);
         if constexpr (residual % 4 == 3) {
             v1 = _mm_setr_epi16(zero, pVect1[0], zero, pVect1[1], zero, pVect1[2], zero,
                                 zero); // SSE2
